Introduce classes to streamline SIMM calculation
Simm2Parameter is supposed to be the interface for all the constants of
the SIMM 2.0 standard. It exposes methods that accept coordinates, such
that the caller does not need to know the specific format it internally
uses.

This, together with immediate result holders -- like WeightedSensitivity
-- allow the calculation to be streamed from the available coordinates,
without too many case distinctions.